Simple static endpoints and docker build working
