+++ Pipfile
@@ -1,0 +1,19 @@
+[[source]]
+url = "https://pypi.org/simple"
+verify_ssl = true
+name = "pypi"
+
+[scripts]
+start = "sh -c 'mh up mongodb && python3 src/server.py'"
+test = "echo 'Not Yet Implemented'"
+stepci = "echo 'Not Yet Implemented'"
+container = "sh -c 'mh down && docker build --tag ghcr.io/agile-learning-institute/mentorhub-curriculum-api:latest . && mh up curriculum-api'"
+
+[packages]
+flask = "*"
+pymongo = "*"
+
+[dev-packages]
+
+[requires]
+python_version = "3.12"
