--- src/flask_app/main.py
@@ -1,41 +1,0 @@
-from flask import Flask, jsonify
-
-app = Flask(__name__)
-
-# GET /api/curriculum/{id}
-@app.route('/api/curriculum/<id>', methods=['GET'])
-def get_or_create_curriculum(id):
-    return jsonify({"hello": "world"})
-
-# PATCH /api/curriculum/{id}
-@app.route('/api/curriculum/<id>', methods=['PATCH'])
-def update_curriculum(id):
-    return jsonify({"hello": "world"})
-
-# POST /api/curriculum/{id}/topic
-@app.route('/api/curriculum/<id>/topic', methods=['POST'])
-def add_topic_to_curriculum(id):
-    return jsonify({"hello": "world"})
-
-# PATCH /api/curriculum/{id}/{topic_id}
-@app.route('/api/curriculum/<id>/<topic_id>', methods=['PATCH'])
-def update_topic_in_curriculum(id, topic_id):
-    return jsonify({"hello": "world"})
-
-# DELETE /api/curriculum/{id}/{topic_id}
-@app.route('/api/curriculum/<id>/<topic_id>', methods=['DELETE'])
-def delete_topic_from_curriculum(id, topic_id):
-    return jsonify({"hello": "world"})
-
-# GET /api/config/
-@app.route('/api/config/', methods=['GET'])
-def get_api_config():
-    return jsonify({"hello": "world"})
-
-# GET /api/health/
-@app.route('/api/health/', methods=['GET'])
-def health_check():
-    return jsonify({"hello": "world"})
-
-if __name__ == '__main__':
-    app.run(debug=True)
